requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
transformers
torch
praw
google-play-scraper
app-store-web-scraper
tenacity
